[PHPStanRules] Remove RequireChildClassGenericTypeRule, already covered by PHPStan level 6
