+++ packages/phpstan-rules/config/symplify-rules.neon
@@ -4,7 +4,6 @@ includes:
 
     # 2. prepared sets
     - nette-rules.neon
-    - generic-rules.neon
     - array-rules.neon
     - code-complexity-rules.neon
     - forbidden-static-rules.neon
