--- packages/phpstan-rules/tests/Rules/RequireChildClassGenericTypeRule/Fixture/MissingGenericType.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Symplify\PHPStanRules\Tests\Rules\RequireChildClassGenericTypeRule\Fixture;
-
-use Symplify\PHPStanRules\Tests\Rules\RequireChildClassGenericTypeRule\Source\AbstractClassWithTemplate;
-
-final class MissingGenericType extends AbstractClassWithTemplate
-{
-
-}
