--- packages/phpstan-rules/tests/Rules/RequireChildClassGenericTypeRule/Source/AbstractClassWithTemplate.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Symplify\PHPStanRules\Tests\Rules\RequireChildClassGenericTypeRule\Source;
-
-/**
- * @template T as stdClass
- */
-abstract class AbstractClassWithTemplate
-{
-}
